feat: moove blockwise coregistration
